move int to float in translateCoordinates helper function

+++ src/main/helpers/helpers.kt
@@ -80,7 +80,7 @@ fun drawRectanglesOnPage (doc : PDDocument, page : PDPage, rectangles : ArrayLis
 
 //Rectangle has origin(0,0) in topLeft corner
 //pdfbox PDRectangle() has origin(0,0) in bottomLeft corner
-fun translateCoordinates(bottomLeftX: Int, bottomLeftY: Int, maxX: Int, maxY: Int): Pair<Int, Int> {
+fun translateCoordinates(bottomLeftX: Float, bottomLeftY: Float, maxX: Float, maxY: Float): Pair<Float, Float> {
     val topLeftX = bottomLeftX
     val topLeftY = maxY - bottomLeftY
     return Pair(topLeftX, topLeftY)
